[jobs,overview] clarify information about merge_auto functionality
We've repeatedly bumped into problems where we have forgotten that regular
dicts don't play well with merge_auto. And this is something we would still
have to repeatedly explain to juniors. So update the documentation reflecting
the fact that not everything is supported, and point people to a place where
they can find the list of things supported.